+++ services/products-service/internal/app/grpcapp/app.go
@@ -61,7 +61,7 @@ func NewApp(log *zap.SugaredLogger, productsInterface products.Products, sagaRes
 
 func interceptorLogger(l *zap.SugaredLogger) logging.Logger {
 	return logging.LoggerFunc(func(ctx context.Context, lvl logging.Level, msg string, fields ...any) {
-		level := zapcore.Level(lvl)
+		level := zapcore.Level(int8(lvl))
 		l.Log(level, msg)
 	})
 }
