+++ services/sso-service/internal/app/grpc/app.go
@@ -50,7 +50,7 @@ func NewApp(log *zap.SugaredLogger, authService authgrpc.Auth, validator validat
 
 func interceptorLogger(l *zap.SugaredLogger) logging.Logger {
 	return logging.LoggerFunc(func(ctx context.Context, lvl logging.Level, msg string, fields ...any) {
-		level := zapcore.Level(lvl)
+		level := zapcore.Level(int8(lvl))
 		l.Log(level, msg)
 	})
 }
